Move useNotificationSubscription to ui-api
