+++ scm-ui/ui-webapp/src/containers/Notifications.tsx
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-import React, { FC, useCallback, useEffect, useState } from "react";
+import React, { FC } from "react";
 import {
   Button,
   Notification as InfoNotification,
@@ -35,7 +35,7 @@ import {
   DateFromNow
 } from "@scm-manager/ui-components";
 import styled from "styled-components";
-import { apiClient, useClearNotifications, useNotifications } from "@scm-manager/ui-api";
+import { useClearNotifications, useNotifications, useNotificationSubscription } from "@scm-manager/ui-api";
 import { Notification, NotificationCollection, Link as LinkType } from "@scm-manager/ui-types";
 import { Link } from "react-router-dom";
 
@@ -152,88 +152,8 @@ const color = (notification: Notification) => {
   return c;
 };
 
-const useSuspendingNotification = (link: string, refetch: () => Promise<NotificationCollection | undefined>) => {
-  const [notifications, setNotifications] = useState<Notification[]>([]);
-  const [disconnectedAt, setDisconnectedAt] = useState<Date>();
-
-  const onVisible = useCallback(() => {
-    return refetch().then(collection => {
-      if (collection) {
-        const received = collection._embedded.notifications.filter(n => {
-          return disconnectedAt && disconnectedAt < new Date(n.createdAt);
-        });
-        if (received.length > 0) {
-          setNotifications(previous => [...previous, ...received]);
-        }
-        setDisconnectedAt(undefined);
-      }
-    });
-  }, [disconnectedAt, refetch]);
-
-  const onHide = useCallback(() => {
-    setDisconnectedAt(new Date());
-    return Promise.resolve();
-  }, []);
-
-  const received = useCallback(
-    (notification: Notification) => {
-      setNotifications(previous => [...previous, notification]);
-      refetch();
-    },
-    [refetch]
-  );
-
-  useEffect(() => {
-    if (link) {
-      let cancel: () => void;
-
-      const disconnect = () => {
-        if (cancel) {
-          cancel();
-        }
-      };
-
-      const connect = () => {
-        disconnect();
-        cancel = apiClient.subscribe(link, {
-          notification: event => {
-            received(JSON.parse(event.data));
-          }
-        });
-      };
-
-      const handleVisibilityChange = () => {
-        if (document.visibilityState === "visible") {
-          onVisible();
-        } else {
-          onHide();
-        }
-      };
-
-      if (document.visibilityState === "visible") {
-        connect();
-      }
-
-      document.addEventListener("visibilitychange", handleVisibilityChange);
-
-      return () => {
-        disconnect();
-        document.removeEventListener("visibilitychange", handleVisibilityChange);
-      };
-    }
-  }, [link, onVisible, onHide, received]);
-
-  return {
-    notifications,
-    remove: (notification: Notification) => {
-      setNotifications([...notifications.filter(n => n !== notification)]);
-    },
-    clear: () => setNotifications([])
-  };
-};
-
 const NotificationSubscription: FC<SubscriptionProps> = ({ link, refetch }) => {
-  const { notifications, remove } = useSuspendingNotification(link, refetch);
+  const { notifications, remove } = useNotificationSubscription(link, refetch);
   return (
     <ToastArea>
       {notifications.map((notification, i) => (
